requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
hydra-core>=1.3.0
mlflow>=2.8.0
dvc[s3]>=3.32.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
albumentations>=1.3.0
tqdm>=4.66.0
python-multipart>=0.0.6
pyyaml>=6.0.1
loguru>=0.7.0
click>=8.1.0
pydantic>=2.4.0
psutil>=5.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
omegaconf>=2.3.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
httpx>=0.24.0

# Code Quality
black>=23.11.0
isort>=5.12.0
ruff>=0.1.6
bandit>=1.7.5
pre-commit>=3.5.0

# Type Checking
mypy>=1.6.0
types-requests>=2.31.0
types-pillow>=10.0.0
types-pyyaml>=6.0.0